requirements.txt
fastapi
uvicorn[standard]
gitpython
PyGithub
openai
langchain
python-dotenv
groq